+++ app/src/main/java/org/android/go/sopt/auth/SignInViewModel.kt
@@ -3,12 +3,11 @@ package org.android.go.sopt.auth
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import kotlinx.coroutines.launch
 import org.android.go.sopt.data.remote.api.MemberServicePool
-import org.android.go.sopt.data.remote.model.RequestSignInDto
+import org.android.go.sopt.data.remote.datasource.AuthRemoteDataSource
 import org.android.go.sopt.data.remote.model.ResponseSignInDto
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 
 class SignInViewModel : ViewModel() {
     private val _signInResult: MutableLiveData<ResponseSignInDto> = MutableLiveData()
@@ -18,30 +17,21 @@ class SignInViewModel : ViewModel() {
     private val _signInMessage: MutableLiveData<String> = MutableLiveData()
     val signInMessage: LiveData<String> = _signInMessage
     private val loginService = MemberServicePool.loginService
+    private val authRemoteDataSource = AuthRemoteDataSource(loginService)
 
     fun signIn(id: String, pw: String) {
-        loginService.signIn(
-            RequestSignInDto(
-                id,
-                pw
-            )
-        ).enqueue(object : Callback<ResponseSignInDto> {
-            override fun onResponse(
-                call: Call<ResponseSignInDto>,
-                response: Response<ResponseSignInDto>
-            ) {
+        viewModelScope.launch {
+            try {
+                val response = authRemoteDataSource.signIn(id, pw)
                 if (response.isSuccessful) {
                     _signInMessage.value = response.body()?.message ?: "로그인에 성공했습니다."
                     _signInResult.value = response.body()
                 } else {
-                    _signInMessage.value = response.body()?.message ?: "로그인에 실패했습니다."
+                    _signInMessage.value = "로그인에 실패했습니다."
                 }
+            } catch (e: Exception) {
+                _signInMessage.value = e.message ?: "로그인 중 오류가 발생했습니다."
             }
-
-            override fun onFailure(call: Call<ResponseSignInDto>, t: Throwable) {
-                _signInMessage.value = t.message ?: "서버와 통신이 원활하지 않습니다."
-            }
-
-        })
+        }
     }
 }
